moves dirs
